fix(locale): rename ja and zh_CN company affix files
