+++ src/locales/ja/company/legal_entity_type.ts
