+++ src/locales/ja/company/name_pattern.ts
@@ -1,4 +1,4 @@
 export default [
-  '{{company.type}}{{person.last_name.generic}}{{company.category}}',
-  '{{person.last_name.generic}}{{company.category}}{{company.type}}',
+  '{{company.legal_entity_type}}{{person.last_name.generic}}{{company.category}}',
+  '{{person.last_name.generic}}{{company.category}}{{company.legal_entity_type}}',
 ];
