+++ src/locales/zh_CN/company/legal_entity_type.ts
