+++ src/locales/zh_CN/company/name_pattern.ts
@@ -1,4 +1,4 @@
 export default [
-  '{{location.city}}{{person.first_name.generic}}{{company.category}}{{company.type}}',
-  '{{location.state}}{{person.first_name.generic}}{{company.category}}{{company.type}}',
+  '{{location.city}}{{person.first_name.generic}}{{company.category}}{{company.legal_entity_type}}',
+  '{{location.state}}{{person.first_name.generic}}{{company.category}}{{company.legal_entity_type}}',
 ];
